refactor: separate final exam components
